# OrderManagement.Api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Order Management API",
    "version": "1.0",
    "description": "Multi-tenant restaurant order management platform for Malaysia",
    "contact": {
      "name": "Order Management Support",
      "email": "support@ordermanagement.local"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5073",
      "description": "Local Development Server (HTTP)"
    },
    {
      "url": "https://localhost:7118",
      "description": "Local Development Server (HTTPS - requires SSL verification disabled)"
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token only"
      }
    },
    "schemas": {
...
      "MenuItemResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "branchId": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440001"
          },
          "name": {
            "type": "string",
            "example": "Dosa"
          },
          "category": {
            "type": "string",
            "example": "South Indian"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 12.50
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
          }
        }
      },
...
      "UpdateMenuItemRequest": {
        "type": "object",
        "required": ["name", "category", "price", "isAvailable"],
        "properties": {
          "name": {
            "type": "string",
            "example": "Dosa Special"
          },
          "category": {
            "type": "string",
            "example": "South Indian"
          },
          "price": {
            "type": "number",
            "format": "decimal",
            "example": 15.00
          },
          "isAvailable": {
            "type": "boolean",
            "example": true
          },
          "imageUrl": {
            "type": "string",
            "nullable": true,
            "example": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
          }
        }
      },
...
      "MenuManagementItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440020"
          },
          "label": {
            "type": "string",
            "example": "Dashboard"
          },
          "icon": {
            "type": "string",
            "nullable": true,
            "example": "dashboard"
          },
          "route": {
            "type": "string",
            "nullable": true,
            "example": "/dashboard"
          },
          "parentId": {
            "type": "string",
            "format": "uuid",
            "nullable": true,
            "example": null
          },
          "displayOrder": {
            "type": "integer",
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "allowedRoles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Admin", "Manager"]
          }
        }
      },
      "UpdateMenuPermissionsRequest": {
        "type": "object",
        "required": ["allowedRoles"],
        "properties": {
          "allowedRoles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Admin", "Manager", "Waiter"]
          }
        }
      },
      "ReorderMenuRequest": {
        "type": "object",
        "required": ["displayOrder"],
        "properties": {
          "displayOrder": {
            "type": "integer",
            "example": 2
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/MenuItems/{id}": {
      "get": {
        "tags": ["Menu Items"],
        "summary": "Get menu item by ID",
        "description": "Retrieve a specific menu item by its ID",
        "operationId": "GetMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Menu item found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItemResponse"
                },
                "example": {
                  "id": "550e8400-e29b-41d4-a716-446655440000",
                  "branchId": "550e8400-e29b-41d4-a716-446655440001",
                  "name": "Dosa",
                  "category": "South Indian",
                  "price": 12.50,
                  "isAvailable": true,
                  "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
                }
              }
            }
          },
          "404": {
            "description": "Menu item not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "put": {
        "tags": ["Menu Items"],
        "summary": "Update menu item",
        "description": "Update an existing menu item",
        "operationId": "UpdateMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemRequest"
              },
              "example": {
                "name": "Dosa Special",
                "category": "South Indian",
                "price": 15.00,
                "isAvailable": true,
                "imageUrl": "https://images.unsplash.com/photo-1612929633736-8c8b2f8b8b8b?w=800"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Menu item updated successfully"
          },
          "404": {
            "description": "Menu item not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": ["Menu Items"],
        "summary": "Delete menu item",
        "description": "Delete a menu item",
        "operationId": "DeleteMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "204": {
            "description": "Menu item deleted successfully"
          },
          "404": {
            "description": "Menu item not found"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/v1/menu-management/items": {
      "get": {
        "tags": ["Menu Management"],
        "summary": "Get all navigation menu items",
        "description": "Get all navigation menu items (SuperAdmin only)",
        "operationId": "GetAllMenuManagementItems",
        "security": [{"Bearer": []}],
        "responses": {
          "200": {
            "description": "List of menu management items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuManagementItemDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - SuperAdmin only"
          }
        }
      }
    },
    "/api/v1/menu-management/items/{id}": {
      "get": {
        "tags": ["Menu Management"],
        "summary": "Get menu management item by ID",
        "description": "Get a specific navigation menu item",
        "operationId": "GetMenuManagementItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Menu management item found"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v1/menu-management/items/{id}/permissions": {
      "put": {
        "tags": ["Menu Management"],
        "summary": "Update menu item permissions",
        "description": "Update role-based permissions for a navigation menu item",
        "operationId": "UpdateMenuPermissions",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuPermissionsRequest"
              },
              "example": {
                "allowedRoles": ["Admin", "Manager"]
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Permissions updated successfully"
          },
          "404": {
            "description": "Menu item not found"
          }
        }
      },
      "get": {
        "tags": ["Menu Management"],
        "summary": "Get menu item permissions",
        "description": "Get role-based permissions for a navigation menu item",
        "operationId": "GetMenuPermissions",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "example": ["Admin", "Manager", "Waiter"]
              }
            }
          },
          "404": {
            "description": "Menu item not found"
          }
        }
      }
    },
    "/api/v1/menu-management/items/{id}/order": {
      "put": {
        "tags": ["Menu Management"],
        "summary": "Update display order",
        "description": "Update the display order of a navigation menu item",
        "operationId": "UpdateDisplayOrder",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderMenuRequest"
              },
              "example": {
                "displayOrder": 2
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Display order updated successfully"
          },
          "404": {
            "description": "Menu item not found"
          }
        }
      }
    },
    "/api/v1/menu-management/items/{id}/toggle": {
      "put": {
        "tags": ["Menu Management"],
        "summary": "Toggle menu item active status",
        "description": "Enable or disable a navigation menu item",
        "operationId": "ToggleMenuItem",
        "security": [{"Bearer": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "boolean"
              },
              "example": true
            }
          }
        },
        "responses": {
          "204": {
            "description": "Menu item toggled successfully"
          },
          "404": {
            "description": "Menu item not found"
          }
        }
      }
    },
...
  },
...
}